requirements.txt
asgiref
crispy-bootstrap4
Django==3.2
django-crispy-forms
django-select2
sqlparse
tzdata
